Add shared auth setup
We can re-use this to log in when testing different sections of the
service across different test files.

The authenticated browser state is written to a file and saved to the
`playwright/.auth` directory for re-use[1].

This also adds an `example.env` file as a template for the environment
variables that need to be set for running locally.

[1]: https://playwright.dev/docs/auth