Make index row selector robust
On the list of programmes, we were clicking on the link in the first row
to move to the programme page. After a dev data refresh, the order of
the programme list changed (we only sort by name in the UI, and names
are not unique). This broke the tests, since they were now navigating to
an unexpected programme

This ensures we're always navigating to the programme we expect by
searching for the index of the correct row, then interpolating that
index into the appropriate selector